refactor!: Remove `Measurement` from Core
