+++ Core/include/Acts/TrackFitting/GainMatrixUpdater.hpp
@@ -9,7 +9,6 @@
 #pragma once
 
 #include "Acts/Definitions/Direction.hpp"
-#include "Acts/EventData/Measurement.hpp"
 #include "Acts/EventData/MeasurementHelpers.hpp"
 #include "Acts/EventData/MultiTrajectory.hpp"
 #include "Acts/Geometry/GeometryContext.hpp"
